Edited all Long/long to Integer/int. Long got edited out from DAO cla…

+++ src/main/java/app/dao/DirectorDAO.java
@@ -3,7 +3,7 @@ package app.dao;
 import app.entities.Director;
 import jakarta.persistence.EntityManagerFactory;
 
-public class DirectorDAO extends GenericDAO<Director, Long> {
+public class DirectorDAO extends GenericDAO<Director, Integer> {
     public DirectorDAO(EntityManagerFactory emf) {
         super(emf, Director.class);
     }
